backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
google-api-python-client>=2.150.0
google-auth>=2.35.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
python-dotenv>=1.0.0